fix: consolidate ZK proof code, align domain separators, add issue references
- client/lib/zk-proof.ts now re-exports from SDK — eliminates duplication
- Aligned verifier.rs domain tags to syncro:payment:* convention (was zkpay-*)
- Added generateAndVerifyProof to SDK proof-generator
- Removed node:buffer import from SDK (browser-safe now)
- Updated pr_body.txt with #835/#933 references

Closes #835
Ref #933
