+++ sdk/src/zk/index.ts
@@ -1,6 +1,7 @@
 export {
   generatePaymentProof,
   verifyPaymentProof,
+  generateAndVerifyProof,
   type ProofBytes,
   type PaymentProofInput,
   type PaymentProofResult,
